+++ src/util/dhcp.rs
@@ -16,9 +16,11 @@
 
 use anyhow::{anyhow, Context, Result};
 use slog_scope::{debug, trace};
+use std::collections::HashMap;
 use std::fs::File;
 use std::path::Path;
 use std::time::Duration;
+use zbus::{dbus_proxy, zvariant};
 
 use super::key_lookup;
 use crate::retry;
@@ -37,12 +39,71 @@ impl DhcpOption {
             .max_backoff(Duration::from_millis(500))
             .max_retries(60)
             .retry(|_| {
+                match self.try_nm() {
+                    Ok(res) => return Ok(res),
+                    Err(e) => trace!("failed querying NetworkManager: {e:#}"),
+                }
                 match self.try_networkd() {
                     Ok(res) => return Ok(res),
                     Err(e) => trace!("failed querying networkd: {e:#}"),
                 }
                 Err(anyhow!("failed to acquire DHCP option"))
             })
+    }
+
+    fn try_nm(&self) -> Result<String> {
+        let key = match *self {
+            Self::DhcpServerId => "dhcp_server_identifier",
+            Self::AzureFabricAddress => "private_245",
+        };
+
+        // We set up everything from scratch on every attempt.  This isn't
+        // super-efficient but is simple and clear.
+        //
+        // We'd like to set both `property` and `object` attributes on the
+        // trait methods, but that fails to compile, so we create proxies by
+        // hand.
+
+        // query NM for active connections
+        let bus = zbus::blocking::Connection::system().context("connecting to D-Bus")?;
+        let nm = NetworkManagerProxyBlocking::new(&bus).context("creating NetworkManager proxy")?;
+        let conn_paths = nm
+            .active_connections()
+            .context("listing active connections")?;
+
+        // walk active connections
+        for conn_path in conn_paths {
+            if conn_path == "/" {
+                continue;
+            }
+            trace!("found NetworkManager connection: {conn_path}");
+            let conn = NMActiveConnectionProxyBlocking::builder(&bus)
+                .path(conn_path)
+                .context("setting connection path")?
+                .build()
+                .context("creating connection proxy")?;
+
+            // get DHCP options
+            let dhcp_path = conn.dhcp4_config().context("getting DHCP config")?;
+            if dhcp_path == "/" {
+                continue;
+            }
+            debug!("checking DHCP config: {dhcp_path}");
+            let dhcp = NMDhcp4ConfigProxyBlocking::builder(&bus)
+                .path(dhcp_path)
+                .context("setting DHCP config path")?
+                .build()
+                .context("creating DHCP config proxy")?;
+            let options = dhcp.options().context("getting DHCP options")?;
+
+            // check for option
+            if let Some(value) = options.get(key) {
+                return value.try_into().context("reading DHCP option as string");
+            }
+        }
+
+        // not found
+        Err(anyhow!("failed to acquire DHCP option {key}"))
     }
 
     fn try_networkd(&self) -> Result<String> {
@@ -75,4 +136,32 @@ impl DhcpOption {
         }
         Err(anyhow!("failed to acquire DHCP option {key}"))
     }
+}
+
+#[dbus_proxy(
+    default_service = "org.freedesktop.NetworkManager",
+    default_path = "/org/freedesktop/NetworkManager",
+    interface = "org.freedesktop.NetworkManager"
+)]
+trait NetworkManager {
+    #[dbus_proxy(property)]
+    fn active_connections(&self) -> zbus::Result<Vec<zvariant::ObjectPath>>;
+}
+
+#[dbus_proxy(
+    default_service = "org.freedesktop.NetworkManager",
+    interface = "org.freedesktop.NetworkManager.Connection.Active"
+)]
+trait NMActiveConnection {
+    #[dbus_proxy(property)]
+    fn dhcp4_config(&self) -> zbus::Result<zvariant::ObjectPath>;
+}
+
+#[dbus_proxy(
+    default_service = "org.freedesktop.NetworkManager",
+    interface = "org.freedesktop.NetworkManager.DHCP4Config"
+)]
+trait NMDhcp4Config {
+    #[dbus_proxy(property)]
+    fn options(&self) -> Result<HashMap<String, zvariant::Value>>;
 }